Extracted the "Remaining time" top card to a separate component
