+++ client/my-sites/plans/ecommerce-trial/ecommerce-trial-banner/style.scss
@@ -1,0 +1,56 @@
+@import "@wordpress/base-styles/breakpoints";
+
+.e-commerce-trial-banner {
+	display: flex;
+	padding: 40px;
+	background-color: rgba(var(--color-neutral-0-rgb), 0.8);
+	box-shadow: none;
+
+	@media (max-width: $break-mobile) {
+		padding: 24px;
+		margin: 0 20px;
+	}
+
+	.e-commerce-trial-banner__content {
+		flex: 1;
+		margin-top: 10px;
+		padding-right: 25%;
+
+		.e-commerce-trial-banner__title {
+			font-size: $font-title-medium;
+			font-weight: 400;
+			color: var(--studio-black);
+			margin: 0;
+
+			@media (max-width: $break-mobile) {
+				font-size: $font-body-small;
+			}
+		}
+
+		.e-commerce-trial-banner__subtitle {
+			font-size: $font-body;
+			font-weight: 400;
+			margin: 15px 0;
+			color: var(--color-neutral-40);
+
+			@media (max-width: $break-mobile) {
+				font-size: $font-body-extra-small;
+				color: var(--color-neutral-80);
+			}
+		}
+	}
+
+	.e-commerce-trial-banner__chart-wrapper {
+		@media (max-width: $break-mobile) {
+			display: none;
+		}
+
+		margin-left: 50px;
+		padding: 0 20px;
+		text-align: center;
+
+		.e-commerce-trial-banner__chart-label {
+			color: var(--color-neutral);
+		}
+	}
+}
