+++ client/my-sites/plans/ecommerce-trial/index.tsx
@@ -1,52 +1,13 @@
 import { Button, Card } from '@automattic/components';
-import { useLocale } from '@automattic/i18n-utils';
 import { useTranslate } from 'i18n-calypso';
-import { useSelector } from 'react-redux';
-import { useLocalizedMoment } from 'calypso/components/localized-moment';
 import BodySectionCssClass from 'calypso/layout/body-section-css-class';
-import {
-	getCurrentPlan,
-	getECommerceTrialDaysLeft,
-	getECommerceTrialExpiration,
-	isECommerceTrialExpired,
-} from 'calypso/state/sites/plans/selectors';
-import { getSelectedSiteId } from 'calypso/state/ui/selectors';
-import DoughnutChart from '../doughnut-chart';
+import ECommerceTrialBanner from './ecommerce-trial-banner';
 import TrialFeatureCard from './trial-feature-card';
 
 import './style.scss';
 
 const ECommerceTrialPlansPage = () => {
-	const selectedSiteId = useSelector( ( state ) => getSelectedSiteId( state ) ) || -1;
-
-	const { currentPlan, eCommerceTrialDaysLeft, isTrialExpired, eCommerceTrialExpiration } =
-		useSelector( ( state ) => ( {
-			currentPlan: getCurrentPlan( state, selectedSiteId ),
-			isTrialExpired: isECommerceTrialExpired( state, selectedSiteId ),
-			eCommerceTrialDaysLeft: Math.round( getECommerceTrialDaysLeft( state, selectedSiteId ) || 0 ),
-			eCommerceTrialExpiration: getECommerceTrialExpiration( state, selectedSiteId ),
-		} ) );
-
-	const locale = useLocale();
-	const moment = useLocalizedMoment();
 	const translate = useTranslate();
-
-	const trialStart = moment( currentPlan?.subscribedDate );
-	const trialEnd = moment( currentPlan?.expiryDate );
-	const trialDuration = trialEnd.diff( trialStart, 'days' );
-
-	/**
-	 * Trial progress from 0 to 1
-	 */
-	const trialProgress = 1 - eCommerceTrialDaysLeft / trialDuration;
-	const eCommerceTrialDaysLeftToDisplay = isTrialExpired ? 0 : eCommerceTrialDaysLeft;
-
-	// moment.js doesn't have a format option to display the long form in a localized way without the year
-	// https://github.com/moment/moment/issues/3341
-	const readableExpirationDate = eCommerceTrialExpiration?.toDate().toLocaleDateString( locale, {
-		month: 'long',
-		day: 'numeric',
-	} );
 
 	const features = [
 		{
@@ -244,44 +205,7 @@ const ECommerceTrialPlansPage = () => {
 		<>
 			<BodySectionCssClass bodyClass={ [ 'is-ecommerce-trial-plan' ] } />
 
-			<Card className="e-commerce-trial-plans__trial-card">
-				<div className="e-commerce-trial-plans__trial-card-content">
-					<p className="e-commerce-trial-plans__card-title">
-						{ translate( 'You’re in a free trial' ) }
-					</p>
-					<p className="e-commerce-trial-plans__card-subtitle">
-						{ isTrialExpired
-							? translate(
-									'Your free trial has expired. Sign up to a plan to unlock new features and keep your store running.'
-							  )
-							: translate(
-									'Your free trial will end in %(daysLeft)d day. Upgrade to a plan by %(expirationdate)s to unlock new features and start selling.',
-									'Your free trial will end in %(daysLeft)d days. Upgrade to a plan by %(expirationdate)s to unlock new features and start selling.',
-									{
-										count: eCommerceTrialDaysLeftToDisplay,
-										args: {
-											daysLeft: eCommerceTrialDaysLeftToDisplay,
-											expirationdate: readableExpirationDate,
-										},
-									}
-							  ) }
-					</p>
-				</div>
-				<div className="e-commerce-trial-plans__chart-wrapper">
-					<DoughnutChart
-						progress={ trialProgress }
-						text={ eCommerceTrialDaysLeftToDisplay?.toString() }
-					/>
-					<br />
-					<span className="e-commerce-trial-plans__chart-label">
-						{ isTrialExpired
-							? translate( 'Your free trial has expired' )
-							: translate( 'day left in trial', 'days left in trial', {
-									count: eCommerceTrialDaysLeftToDisplay,
-							  } ) }
-					</span>
-				</div>
-			</Card>
+			<ECommerceTrialBanner />
 
 			<Card className="e-commerce-trial-plans__price-card">
 				<div className="e-commerce-trial-plans__price-card-text">
