+++ client/my-sites/plans/ecommerce-trial/style.scss
@@ -9,60 +9,6 @@ body.is-section-plans.is-ecommerce-trial-plan {
 		box-shadow: inset 0 -1px 0 rgba(0, 0, 0, 0.05);
 	}
 
-	.e-commerce-trial-plans__trial-card {
-		display: flex;
-		padding: 40px;
-		background-color: rgba(var(--color-neutral-0-rgb), 0.8);
-		box-shadow: none;
-
-		@media (max-width: $break-mobile) {
-			padding: 24px;
-			margin: 0 20px;
-		}
-
-		.e-commerce-trial-plans__trial-card-content {
-			flex: 1;
-			margin-top: 10px;
-
-			.e-commerce-trial-plans__card-title {
-				font-size: $font-title-medium;
-				font-weight: 400;
-				color: var(--studio-black);
-				margin: 0;
-
-				@media (max-width: $break-mobile) {
-					font-size: $font-body-small;
-				}
-			}
-
-			.e-commerce-trial-plans__card-subtitle {
-				font-size: $font-body;
-				font-weight: 400;
-				margin: 15px 0;
-				color: var(--color-neutral-40);
-
-				@media (max-width: $break-mobile) {
-					font-size: $font-body-extra-small;
-					color: var(--color-neutral-80);
-				}
-			}
-		}
-
-		.e-commerce-trial-plans__chart-wrapper {
-			@media (max-width: $break-mobile) {
-				display: none;
-			}
-
-			margin-left: 50px;
-			padding: 0 20px;
-			text-align: center;
-
-			.e-commerce-trial-plans__chart-label {
-				color: var(--color-neutral);
-			}
-		}
-	}
-
 	.e-commerce-trial-plans__price-card {
 		display: flex;
 		flex-wrap: wrap;
